Validate file copy/removal over dbus for desktopd

+++ dbp-run-vala/dbp-desktopd.vala
@@ -65,10 +65,14 @@ string desktop_path(string full_path) {
 }
 
 void add_meta(string path) {
+	string dest_path = desktop_path(path);;
 	try {
 		var src = File.new_for_path(path);
-		var dest = File.new_for_path(desktop_path(path));
-		src.copy(dest, FileCopyFlags.NONE);
+		var dest = File.new_for_path(dest_path);
+		if (dest_path.has_prefix("__dbp__") && dest_path.has_suffix(".desktop"))
+			src.copy(dest, FileCopyFlags.NONE);
+		else
+			stderr.printf(_("Warning: Requested to copy a file that isn't a DBP .desktop file\n"));
 	} catch (Error e) {
 		stderr.printf("Error: %s\n", e.message);
 		return;
@@ -77,7 +81,13 @@ void add_meta(string path) {
 }
 
 void remove_meta(string path) {
-	GLib.FileUtils.unlink(desktop_path(path));
+	string dest;
+
+	dest = desktop_path(path);
+	if (dest.has_prefix("__dbp__") && dest.has_suffix(".desktop"))
+		GLib.FileUtils.unlink(desktop_path(path));
+	else
+		stderr.printf("Warning: Requested to remove a file that isn't a DBP .desktop file\n");
 	return;
 }
